+++ packages/core/src/benchmark/index.js
@@ -1,5 +1,0 @@
-/**
- * 评测模块导出
- */
-export { SWEBenchEngine, scanSWEBenchData, } from './SWEBenchEngine.js';
-//# sourceMappingURL=index.js.map