+++ packages/core/src/im/index.js
@@ -1,9 +1,0 @@
-/**
- * IM 适配器系统统一导出
- */
-export { BaseIMAdapter } from './BaseIMAdapter.js';
-export { TelegramAdapter } from './TelegramAdapter.js';
-export { FeishuAdapter } from './FeishuAdapter.js';
-export { WeChatAdapter } from './WeChatAdapter.js';
-export { IMManager } from './IMManager.js';
-//# sourceMappingURL=index.js.map