+++ packages/core/src/plugins/index.js
@@ -1,5 +1,0 @@
-// 管理器
-export { PluginManager, getPluginManager, resetPluginManager } from './PluginManager.js';
-// 内置技能
-export { CodeReviewSkill, UnitTestSkill, CodeExplainSkill, GenerateDocSkill, RefactorSkill, DebugSkill, BUILTIN_SKILLS, getSkillsByTag, getSkillByName, } from './BuiltinSkills.js';
-//# sourceMappingURL=index.js.map